fix usage of interface + only report for command
