+++ src/commands/context/report.ts
@@ -4,7 +4,7 @@ import {
   PermissionFlagsBits,
 } from 'discord.js';
 import { ContextMenuCommand } from '../../types';
-import reportMessage from '../../report';
+import { isStaff } from '../../utils';
 
 /**
  * Report message command
@@ -12,6 +12,8 @@ import reportMessage from '../../report';
  * Logs a message in the mod log channel. If a normal user uses the command it will ping the mods
  */
 
+// We will keep a memory cache of warned messages to avoid showing multiple warnings
+const warnedMessageIds: string[] = [];
 
 export const command: ContextMenuCommand = {
   data: new ContextMenuCommandBuilder()
@@ -25,6 +27,20 @@ export const command: ContextMenuCommand = {
 
     if (!guild) return;
 
+    const sendSuccessMessage = () => {
+      interaction.reply({
+        content:
+          'Thanks, the message has been reported and the moderators have been notified',
+        ephemeral: true,
+      });
+    };
+
+    if (warnedMessageIds.includes(targetMessage.id)) {
+      // Send the success message anyway so they know the mods have been notified
+      sendSuccessMessage();
+      return;
+    }
+
     if (targetMessage.author.id === user.id) {
       // Stop this way of pinging mods for code help plz
       interaction.reply({
@@ -34,20 +50,72 @@ export const command: ContextMenuCommand = {
       return;
     }
 
-    const reportResult = await reportMessage(client, guild, targetMessage, user, true)
+    const channel = client.channels.cache.get(process.env.MOD_LOG_CHANNEL_ID);
 
-    if (reportResult === true) {
-      interaction.reply({
-        content:
-          'Thanks, the message has been reported and the moderators have been notified',
-        ephemeral: true,
-      });
+    if (!channel || !channel.isTextBased()) {
+      console.error(
+        `No mod-log channel found (using the ID ${process.env.MOD_LOG_CHANNEL_ID})!`
+      );
 
-    } else {
       interaction.reply({
         content: 'Something went wrong, please try again later',
         ephemeral: true,
       });
+
+      return;
     }
+
+    const author = await guild.members.fetch(targetMessage.author.id);
+    const userGuildMember = await guild.members.fetch(user.id);
+    const isUserStaff = isStaff(userGuildMember);
+
+    channel.send({
+      content: !isUserStaff
+        ? `<@&${process.env.MODERATOR_ROLE_ID}>`
+        : undefined,
+      embeds: [
+        {
+          title: '⚠️ Message Reported',
+          description: '```' + targetMessage.content + '```',
+          color: 16098851,
+          author: {
+            name: author?.displayName ?? 'Unknown user',
+            icon_url: author?.displayAvatarURL(),
+          },
+          fields: [
+            {
+              name: 'Author',
+              value: `<@${targetMessage.author.id}>`,
+              inline: true,
+            },
+            {
+              name: 'Channel',
+              value: `<#${targetMessage.channelId}>`,
+              inline: true,
+            },
+            {
+              name: 'Jump to message',
+              value: `[Click here](${targetMessage.url})`,
+              inline: true,
+            },
+          ],
+          footer: {
+            icon_url: userGuildMember.displayAvatarURL(),
+            text: `Reported by ${userGuildMember.displayName}`,
+          },
+        },
+      ],
+    });
+
+    if (isUserStaff) {
+      interaction.reply({
+        content: 'Message logged in the mod channel',
+        ephemeral: true,
+      });
+    } else {
+      sendSuccessMessage();
+    }
+
+    warnedMessageIds.push(targetMessage.id);
   },
 };
